Correct `gutter_foreground_highlight`'s color
It turned out gutter_foreground_highlight color is blended with the
foreground instead of `background` like `gutter_foreground`
https://github.com/sublimehq/sublime_text/issues/5166#issuecomment-1007219048 .
So the value of `color(var(--foreground) a(0.6))` wouldn't yield the
correct color.
Logically `color(var(--background) a(0.4))` would be the correct color
but it wasn't exactly the same as atom's.
So I had to hard-code the correct colors.