Fix overlay shadow artifacts
The shadow assets exported with Figma contain artifacts.
The ones exported with Inkscape from an svg exported from figma don't
and the result matches atom.
The issue is that, I can't make a transparent rect with shadow in
Inkscape, I had to export it from figma and inkscape as png and copy
the rect region from figma's onto inkscape's exported file using GIMP!