Crispier buttons that match atom
* Exporting `svg` to `png` using Inkscape, especially for assets with
rounded corners, produced a different result than expected.
Atom being an electron app uses `skia` as a rendering engine,
while Inkscape uses `cairo`, the latter arguably produces smoother
corners but using more colors, so assuming that for a `3px` rounded
corner the `layer#.inner_margin` should be set to `3` is wrong in this
case because there are non transparent pixels outside of the `3x3`
corner that get stretched producing bad result.
The solution was to export using figma which uses `skia`.
* Removed `border_middle.png`, `border_middle_pressed.png` and
`button_middle.png` assets.
* The buttons background linear gradient varies vertically and not
horizontally, changed the assets to reflect that.
* Changed assets to provide only the needed data.